+++ test/html-equal.test.js
@@ -1,0 +1,16 @@
+const test = QUnit.test;
+
+test('normalizes whitespace', assert => {
+    const html = /*html*/ `
+            <div>
+                <span>message</span>
+            </div>
+    `;
+    const expected = /*html*/ `
+        <div>
+                <span>message</span>
+        </div>
+    `;
+
+    assert.htmlEqual(html, expected);
+});
